Wrote source xml files
Created loop to go through every pk ID, and with these got all the information for every file. There is an 'r' and a few 's's that were created for some reason, but once we're able to catch errors this should be fully functional. Now to do the same for every other writer.